change type Item.quantity from integer to double

+++ Spender/app/src/main/java/com/project/spender/fns/api/Item.java
@@ -19,7 +19,7 @@ public class Item {
 
     @SerializedName("quantity")
     @Expose
-    public Integer quantity;
+    public Double quantity;
 
     @SerializedName("nds10")
     @Expose
